+++ src/main/java/subway/service/LineService.java
@@ -1,7 +1,9 @@
-package subway.domain;
+package subway.service;
 
 import java.util.Arrays;
 import java.util.List;
+import subway.domain.Line;
+import subway.domain.LineRepository;
 
 /*
  * 지하철 노선에 관련된 기능을 관리하는 클래스
